ar/karajnaproject/requirements.txt
Django==2.0.6
django-filter==0.13.0
